pyproject.toml
[project]
name = "twitter-emotion-classification"
version = "0.1.0"
description = "BERT-based emotion classification for Twitter messages"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.13"
dependencies = [
    "datasets>=2.0.0",
    "imbalanced-learn>=0.8.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "lightning>=2.6.1",
    "matplotlib>=3.3.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "scikit-learn>=0.24.0",
    "seaborn>=0.11.0",
    "tensorboard>=2.10.0",
    "torch==2.7",
    "torchmetrics>=1.0.0",
    "tqdm>=4.60.0",
    "transformers>=4.30.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
]
